fix: enhance image export handling
